requirements.txt
boto==2.46.1
dj-database-url==0.4.2
Django==1.10.6
django-forms-bootstrap==3.0.1
django-gravatar2==1.4.0
django-storages==1.5.2
djangorestframework==3.6.2
gunicorn==19.7.1
olefile==0.44
Pillow==4.0.0
psycopg2==2.7.1
requests==2.13.0
stripe==1.51.0
